requirements.txt
asgiref==3.9.0
Django==5.2.4
django-ckeditor==6.7.3
django-js-asset==3.1.2
sqlparse==0.5.3
tzdata==2025.2
